Tokenizer, parser, interpreter - final

+++ zadanie3/src/main/java/pl/edu/agh/amwj/value/types/IntegerValue.java
@@ -4,12 +4,12 @@ package pl.edu.agh.amwj.value.types;
  * Created by Kurtzz on 2016-11-10.
  */
 public class IntegerValue implements Value {
-    private final double value;
+    private final int value;
 
-    public IntegerValue(double value) {
+    public IntegerValue(int value) {
         this.value = value;
     }
-    @Override public String toString() { return Double.toString(value); }
+    @Override public String toString() { return Integer.toString(value); }
     public double toNumber() { return value; }
 
     public Value evaluate() { return this; }
